Fixed an error on arm64
On the arm64 the comparison with NSNotFound will fail, because it's no longer in the range of int

+++ PopoverView/PopoverView.m
@@ -797,7 +797,7 @@
 
 - (void)didTapButton:(UIButton *)sender
 {
-    int index = [subviewsArray indexOfObject:sender];
+    NSUInteger index = [subviewsArray indexOfObject:sender];
     
     if (index == NSNotFound) {
         return;
